requirements.txt
fastapi
uvicorn
cobra
pyyaml
